#Cleanup main.py, seperate into multiple files
