+++ location.py
@@ -1,0 +1,32 @@
+import struct
+from geopy.geocoders import GoogleV3
+
+COORDS_LATITUDE = 0
+COORDS_LONGITUDE = 0
+COORDS_ALTITUDE = 0
+
+def f2i(float):
+  return struct.unpack('<Q', struct.pack('<d', float))[0]
+
+def f2h(float):
+  return hex(struct.unpack('<Q', struct.pack('<d', float))[0])
+
+def h2f(hex):
+  return struct.unpack('<d', struct.pack('<Q', int(hex,16)))[0]
+
+def set_location(location_name):
+    geolocator = GoogleV3()
+    loc = geolocator.geocode(location_name)
+
+    print('[!] Your given location: {}'.format(loc.address))
+    print('[!] lat/long/alt: {} {} {}'.format(loc.latitude, loc.longitude, loc.altitude))
+    set_location_coords(loc.latitude, loc.longitude, loc.altitude)
+
+def set_location_coords(lat, long, alt):
+    global COORDS_LATITUDE, COORDS_LONGITUDE, COORDS_ALTITUDE
+    COORDS_LATITUDE = f2i(lat)
+    COORDS_LONGITUDE = f2i(long)
+    COORDS_ALTITUDE = f2i(alt)
+
+def get_location_coords():
+    return (COORDS_LATITUDE, COORDS_LONGITUDE, COORDS_ALTITUDE)
