+++ login.py
@@ -1,0 +1,51 @@
+from gpsoauth import perform_master_login, perform_oauth
+
+ANDROID_ID = '9774d56d682e549c'
+SERVICE= 'audience:server:client_id:848232511240-7so421jotr2609rmqakceuu1luuq0ptb.apps.googleusercontent.com'
+APP = 'com.nianticlabs.pokemongo'
+CLIENT_SIG = '321187995bc7cdc2b5fc91b11a96e2baa8602c62'
+
+def login_google(username, password):
+    print('[!] Google login for: {}'.format(username))
+    r1 = perform_master_login(username, password, ANDROID_ID)
+    r2 = perform_oauth(username, r1.get('Token', ''), ANDROID_ID, SERVICE, APP,
+        CLIENT_SIG)
+
+    return r2.get('Auth') # access token
+
+def login_ptc(username, password):
+    print('[!] PTC login for: {}'.format(username))
+    head = {'User-Agent': 'niantic'}
+    r = SESSION.get(LOGIN_URL, headers=head)
+    jdata = json.loads(r.content)
+    data = {
+        'lt': jdata['lt'],
+        'execution': jdata['execution'],
+        '_eventId': 'submit',
+        'username': username,
+        'password': password,
+    }
+    r1 = SESSION.post(LOGIN_URL, data=data, headers=head)
+
+    ticket = None
+    try:
+        ticket = re.sub('.*ticket=', '', r1.history[0].headers['Location'])
+    except Exception,e:
+        if DEBUG:
+            print(r1.json()['errors'][0])
+        return None
+
+    data1 = {
+        'client_id': 'mobile-app_pokemon-go',
+        'redirect_uri': 'https://www.nianticlabs.com/pokemongo/error',
+        'client_secret': PTC_CLIENT_SECRET,
+        'grant_type': 'refresh_token',
+        'code': ticket,
+    }
+    r2 = SESSION.post(LOGIN_OAUTH, data=data1)
+    access_token = re.sub('&expires.*', '', r2.content)
+    access_token = re.sub('.*access_token=', '', access_token)
+
+    return access_token
+
+
