+++ settings.py
@@ -1,0 +1,3 @@
+#Global settings, call init() from main function
+def init():
+    global debug
